Close contours of edge cells.
For cells that intersect an edge, find_contours() leaves the contour open. This was breaking skeletonization, because the projection from the voronoi skeleton does not intersect the contour, and find_segment_intersection() was returning a NaN, which was breaking evenly_distribute_contour_points().

We should really be catching and ignoring edge cells by default, in which case this error condition should never arise. But for kymographs, we want to be able to pass a flag to keep edge cells. They will need to be treated differently, probably cropped perpendicular to their central axis as close to the edge as possible.