Include the fork node on all emanating branches.
This resolves a bug in the graph walker, where every fork node is only added to one of its emanating branches. Because of this, the skeleton was missing one vertex.

Solution: Rather than pushing single nodes onto a stack at each fork node, push pairs of nodes which include the fork node and the next node along each emanating branch.