Handle edge case in contour point redistribution.
In evenly_distribute_contour_points() I was occasionally getting:
RuntimeWarning: invalid value encountered in divide
This was caused by a divide by zero in the calculation of alpha.

If any point of a contour remains perfectly still between updates,
the difference in alpha's numerator and denominator terms will be
zero at that contour coordinate. Since this is a zero-over-zero
situation, we want the resulting value of alpha to be zero anyway.
So, we can just catch any zero in the denominator, skip division,
and set the resulting position in alpha to zero.